ref: Create a dedicated types directory
