+++ src/screens/TabOneScreen.tsx
@@ -2,7 +2,7 @@ import { StyleSheet } from "react-native";
 
 import EditScreenInfo from "../components/EditScreenInfo";
 import { Text, View } from "../components/Themed";
-import { RootTabScreenProps } from "../../types";
+import { RootTabScreenProps } from "../types/navigation/types";
 
 export default function TabOneScreen({
   navigation,
